small class fixes
