--- src/main/java/com/sdt/KingGame/Main.java
@@ -1,6 +1,0 @@
-public class Main {
-    public static void main(String[] args) {
-        Deck deck = new Deck();
-        System.out.println(deck);
-    }
-}
